refactor(routes): extract approval SSE state into api/route_approvals.py (#1907)
